+++ semana19/aula1/jest-template/package.json
@@ -11,10 +11,13 @@
   "dependencies": {
     "@types/express": "^4.17.7",
     "@types/jest": "^25.2.3",
+    "@types/knex": "^0.16.1",
     "@types/node": "^14.0.6",
     "dotenv": "^8.2.0",
     "express": "^4.17.0",
     "jest": "^26.0.1",
+    "knex": "^0.21.2",
+    "mysql": "^2.18.1",
     "ts-jest": "^26.1.0",
     "ts-node-dev": "^1.0.0-pre.44",
     "typescript": "^3.9.3"
