+++ semana19/aula1/jest-template/src/BaseDatabase.ts
