--- semana19/aula1/jest-template/src/services/Authenticator.ts
@@ -1,34 +1,0 @@
-import * as jwt from "jsonwebtoken";
-
-export class Authenticator {
-  
-  private static EXPIRES_IN = "30min";
-  public generateToken(input: AuthenticationData): string {
-    const token = jwt.sign(
-      {
-        id: input.id,
-      },
-      process.env.JWT_KEY as string,
-      {
-        expiresIn: Authenticator.EXPIRES_IN,
-      }
-    );
-    return token;
-  }
-
-  public verifyToken(token: string): AuthenticationData {
-    const payload = jwt.verify(token, process.env.JWT_KEY as string) as any;
-    const result = {
-      id: payload.id,
-    };
-    return result;
-  }
-  public getData = (token: string): AuthenticationData => {
-    const payload = jwt.verify(token, process.env.JWT_KEY as string) as any;
-    return payload;
-  };
-}
-
-interface AuthenticationData {
-  id: string;
-}
