--- semana19/aula1/jest-template/src/services/IdGenerator.ts
@@ -1,7 +1,0 @@
-import { v4 } from "uuid";
-
-export class IdGenerator {
-  public generate(): string {
-    return v4();
-  }
-}
